refactor: split type model module
